feat(TDD cycle: Red stage): Add failing test to print table.
This new test describe a desired behavior for the `MultiplicationTable`
class: "It should print the multiplication table by five".

This test "should fail!", because the code to make it pass hasn't been
written yet (`printTable` method doesn't exist yet).

Issue: "The method printTable(int) is undefined for the type
MultiplicatonTable"