feat(TDD cycle: Red stage): Add failing test to generate string table.
This new test describe a desired behavior for the `MultiplicationTable`
class: "It should return the multiplication table for a number as a
string".

This test "should fail!", because the code to make it pass hasn't been
written yet (`generateTableString` method doesn't exist yet).

Refactor Objective: Apply 'Separation of Concerns'
Separate 'the logic of generate the output' from 'the act of print it
to the console'. This makes the code more modular, flexible, and easier
to test.